requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
Werkzeug
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart
websockets
aiofiles